refactor(fast_io): decompose iocp/disk_batch.rs into submodules (SPL-26 #2454)
Split the 1263-line iocp/disk_batch.rs into a directory module with
focused submodules, each well under the 650-line cap:

- mod.rs (386): IocpDiskBatch struct, ActiveFile, lifecycle methods,
  Write and Drop impls
- buffer.rs (84): BatchBuffer arena enum (Vec / PageAligned), bounce-copy
  telemetry counter and public accessors
- writer.rs (255): WriteFile dispatch, overlapped-handle reopen/close,
  batched submission and short-write resubmission
- completion.rs (183): GetQueuedCompletionStatusEx drain, NTSTATUS to DOS
  error mapping, test-only fault injector
- tests.rs (425): every existing unit test, unchanged behavioural coverage

Zero public API change: iocp/mod.rs still re-exports IocpDiskBatch,
bounce_copies_avoided, and the doc-hidden test hooks via the same paths.
All unsafe blocks retain their SAFETY comments. No clippy waivers added.
