+++ crates/fast_io/src/iocp/disk_batch/completion.rs
@@ -1,0 +1,183 @@
+//! Completion port draining, NTSTATUS mapping, and the test-only
+//! `WriteFile` fault injector.
+//!
+//! The drain loop reaps overlapped completions from the single port owned by
+//! [`super::IocpDiskBatch`] using `GetQueuedCompletionStatusEx`. Errors
+//! surfaced through the `OVERLAPPED::Internal` NTSTATUS field are translated
+//! into Win32 DOS error codes by [`ntstatus_to_dos_error`] so they round-trip
+//! through `io::Error::from_raw_os_error` like any synchronous Win32 failure.
+//!
+//! The fault injector is a `#[doc(hidden)]` hook used by the disk-full
+//! simulation test to drive deterministic `ERROR_DISK_FULL` coverage without
+//! provisioning a limited-capacity volume.
+
+use std::io;
+use std::sync::atomic::{AtomicI32, AtomicUsize, Ordering};
+
+use windows_sys::Win32::Foundation::{ERROR_HANDLE_EOF, FALSE, WAIT_TIMEOUT};
+use windows_sys::Win32::System::IO::{GetQueuedCompletionStatusEx, OVERLAPPED_ENTRY};
+
+use crate::iocp::completion_port::CompletionPort;
+
+/// Maximum entries dequeued by a single `GetQueuedCompletionStatusEx` call.
+///
+/// Matches the io_uring side's CQE batch sizing so both backends use the
+/// same drain granularity. Kept fixed (not CPU-scaled) because the disk
+/// batch drains exactly the in-flight cohort capped by
+/// `IocpConfig::concurrent_ops`, which is already auto-sized by
+/// `super::config::default_concurrent_ops`. The clamp ceiling
+/// (`super::config::MAX_CONCURRENT_OPS`) is intentionally aligned with this
+/// constant so a single drain call can reap the entire cohort.
+const COMPLETION_DRAIN_BATCH: usize = 64;
+
+/// Wait timeout for completion drains, in milliseconds. The disk batch
+/// always knows how many completions are outstanding so it waits
+/// indefinitely (`u32::MAX`) until every submitted write has been reaped.
+const DRAIN_TIMEOUT_MS: u32 = u32::MAX;
+
+/// Countdown for the test-only `WriteFile` fault injector. When non-zero,
+/// each submission decrements the counter; once it hits zero the next
+/// submission returns the OS error code stashed in
+/// [`FAULT_INJECT_ERROR_CODE`] instead of issuing `WriteFile`. The hook is
+/// dormant by default (counter == 0, error == 0) and adds a single relaxed
+/// load to the submit hot path - negligible compared to the syscall it
+/// guards.
+#[doc(hidden)]
+static FAULT_INJECT_COUNTDOWN: AtomicUsize = AtomicUsize::new(0);
+
+/// Win32 error code returned by the next fault-injected submission. Paired
+/// with [`FAULT_INJECT_COUNTDOWN`]; see [`inject_next_write_error_for_test`].
+#[doc(hidden)]
+static FAULT_INJECT_ERROR_CODE: AtomicI32 = AtomicI32::new(0);
+
+/// Arms the test-only `WriteFile` fault injector so the `nth` (1-based)
+/// upcoming submission inside [`super::IocpDiskBatch`] returns `os_error`
+/// instead of dispatching to the kernel. Subsequent submissions proceed
+/// normally after the single fault fires.
+///
+/// This is a test hook used by `crates/fast_io/tests/iocp_disk_full_simulation.rs`
+/// to drive deterministic ERROR_DISK_FULL coverage without provisioning a
+/// limited-capacity volume. Production code must never call it; the hook is
+/// `#[doc(hidden)]` and excluded from the public API surface.
+#[doc(hidden)]
+pub fn inject_next_write_error_for_test(nth: usize, os_error: i32) {
+    FAULT_INJECT_ERROR_CODE.store(os_error, Ordering::SeqCst);
+    FAULT_INJECT_COUNTDOWN.store(nth, Ordering::SeqCst);
+}
+
+/// Clears any pending fault-injection state. Tests should call this at the
+/// end of every case so a leftover countdown from a panicking test does not
+/// leak into a sibling test that shares the same process.
+#[doc(hidden)]
+pub fn clear_injected_write_error_for_test() {
+    FAULT_INJECT_COUNTDOWN.store(0, Ordering::SeqCst);
+    FAULT_INJECT_ERROR_CODE.store(0, Ordering::SeqCst);
+}
+
+/// Returns `Some(os_error)` if the current submission should be faulted, or
+/// `None` to dispatch normally. Atomically decrements the countdown so only
+/// one submission per arm fires.
+pub(super) fn take_injected_write_error() -> Option<i32> {
+    // Cheap relaxed fast path: when no test has armed the hook (the common
+    // case in production) we do a single load and return.
+    if FAULT_INJECT_COUNTDOWN.load(Ordering::Relaxed) == 0 {
+        return None;
+    }
+    // Slow path uses SeqCst to interleave correctly with the arming store.
+    let prev = FAULT_INJECT_COUNTDOWN
+        .fetch_update(Ordering::SeqCst, Ordering::SeqCst, |n| match n {
+            0 => None,
+            n => Some(n - 1),
+        })
+        .ok()?;
+    if prev == 1 {
+        let code = FAULT_INJECT_ERROR_CODE.swap(0, Ordering::SeqCst);
+        if code != 0 {
+            return Some(code);
+        }
+    }
+    None
+}
+
+/// Drains up to `max` completion entries from the port using
+/// `GetQueuedCompletionStatusEx` and returns
+/// `(overlapped_address, bytes_transferred)` pairs.
+pub(super) fn drain_completions(
+    port: &CompletionPort,
+    max: usize,
+) -> io::Result<Vec<(usize, usize)>> {
+    let batch = max.clamp(1, COMPLETION_DRAIN_BATCH);
+    let mut entries: Vec<OVERLAPPED_ENTRY> = vec![zeroed_entry(); batch];
+
+    loop {
+        let mut removed: u32 = 0;
+        // SAFETY: `port.handle()` is owned by `port` and lives for the
+        // duration of the call; `entries` backs `batch` slots.
+        #[allow(unsafe_code)]
+        let ok = unsafe {
+            GetQueuedCompletionStatusEx(
+                port.handle(),
+                entries.as_mut_ptr(),
+                batch as u32,
+                &mut removed,
+                DRAIN_TIMEOUT_MS,
+                FALSE,
+            )
+        };
+
+        if ok == FALSE {
+            let err = io::Error::last_os_error();
+            // Spurious wake without entries: retry.
+            if matches!(err.raw_os_error(), Some(c) if c as u32 == WAIT_TIMEOUT) {
+                continue;
+            }
+            return Err(err);
+        }
+
+        let mut out = Vec::with_capacity(removed as usize);
+        for entry in entries.iter().take(removed as usize) {
+            let overlapped_ptr = entry.lpOverlapped;
+            if overlapped_ptr.is_null() {
+                continue;
+            }
+            // SAFETY: entry.lpOverlapped points at the OVERLAPPED structure
+            // we submitted; the surrounding pinned op is still alive in
+            // the in-flight queue, so reading the Internal field is sound.
+            #[allow(unsafe_code)]
+            let internal = unsafe { (*overlapped_ptr).Internal };
+            if internal != 0 {
+                let dos_error = ntstatus_to_dos_error(internal as u32);
+                if dos_error == ERROR_HANDLE_EOF {
+                    return Err(io::Error::from(io::ErrorKind::UnexpectedEof));
+                }
+                return Err(io::Error::from_raw_os_error(dos_error as i32));
+            }
+            out.push((
+                overlapped_ptr as usize,
+                entry.dwNumberOfBytesTransferred as usize,
+            ));
+        }
+        return Ok(out);
+    }
+}
+
+/// Translates the small set of NTSTATUS codes that overlapped file I/O can
+/// produce into Win32 DOS error codes.
+fn ntstatus_to_dos_error(status: u32) -> u32 {
+    match status {
+        0xC000_0011 => ERROR_HANDLE_EOF, // STATUS_END_OF_FILE
+        0xC000_0120 => 995,              // STATUS_CANCELLED
+        0xC000_009A => 1450,             // STATUS_INSUFFICIENT_RESOURCES
+        0xC000_00B5 => 121,              // STATUS_IO_TIMEOUT
+        other => other,
+    }
+}
+
+/// Constructs a zeroed `OVERLAPPED_ENTRY` for batch dequeues.
+fn zeroed_entry() -> OVERLAPPED_ENTRY {
+    // SAFETY: OVERLAPPED_ENTRY is plain-old-data and valid when zeroed.
+    #[allow(unsafe_code)]
+    unsafe {
+        std::mem::zeroed()
+    }
+}
